seq replaces the dependency on dash.
Mostly, this change is a like-for-like replacement, with two exceptions:
seq.el is missing a tree-map function which is replaced by recursion;
and, `m-buffer-partition-by-marker' has been altered as it was making
implicit use of the order of evaluation of group-by which,
unfortunately, differs between dash and seq.

Version number has been bumped to Emacs-25 (since seq-24.el lacks the
`seqp' function). Travis has been updated to reflect this.

Moved to assess-discover for tests because it made life easier.